Adding 3D scatterplot HTML file
Three-dimensional scatterplot for cell cluster analysis. Labeled in red are cells that did not migrate from baseline cluster 0 to natural_odor cluster 1. Hovering over a data point shows the main values for the three variables of interest (Frequency, Synchro and Normalized AUC) as well as the cell's ID (Animal #, ROI#).